+++ processor/batchprocessor/batch_processor_test.go
@@ -31,8 +31,9 @@ import (
 	"go.opentelemetry.io/collector/consumer"
 	"go.opentelemetry.io/collector/consumer/consumertest"
 	"go.opentelemetry.io/collector/internal/testdata"
-	"go.opentelemetry.io/collector/model/otlp"
-	"go.opentelemetry.io/collector/model/pdata"
+	"go.opentelemetry.io/collector/pdata/plog"
+	"go.opentelemetry.io/collector/pdata/pmetric"
+	"go.opentelemetry.io/collector/pdata/ptrace"
 )
 
 func TestBatchProcessorSpansDelivered(t *testing.T) {
@@ -46,7 +47,7 @@ func TestBatchProcessorSpansDelivered(t *testing.T) {
 
 	requestCount := 1000
 	spansPerRequest := 100
-	traceDataSlice := make([]pdata.Traces, 0, requestCount)
+	traceDataSlice := make([]ptrace.Traces, 0, requestCount)
 	for requestNum := 0; requestNum < requestCount; requestNum++ {
 		td := testdata.GenerateTracesManySpansSameResource(spansPerRequest)
 		spans := td.ResourceSpans().At(0).ScopeSpans().At(0).Spans()
@@ -58,7 +59,7 @@ func TestBatchProcessorSpansDelivered(t *testing.T) {
 	}
 
 	// Added to test logic that check for empty resources.
-	td := pdata.NewTraces()
+	td := ptrace.NewTraces()
 	assert.NoError(t, batcher.ConsumeTraces(context.Background(), td))
 
 	require.NoError(t, batcher.Shutdown(context.Background()))
@@ -98,7 +99,7 @@ func TestBatchProcessorSpansDeliveredEnforceBatchSize(t *testing.T) {
 	}
 
 	// Added to test logic that check for empty resources.
-	td := pdata.NewTraces()
+	td := ptrace.NewTraces()
 	require.NoError(t, batcher.ConsumeTraces(context.Background(), td))
 
 	// wait for all spans to be reported
@@ -120,7 +121,7 @@ func TestBatchProcessorSpansDeliveredEnforceBatchSize(t *testing.T) {
 }
 
 func TestBatchProcessorSentBySize(t *testing.T) {
-	sizer := otlp.NewProtobufTracesMarshaler().(pdata.TracesSizer)
+	sizer := ptrace.NewProtoMarshaler().(ptrace.Sizer)
 	views := MetricViews()
 	require.NoError(t, view.Register(views...))
 	defer view.Unregister(views...)
@@ -276,7 +277,7 @@ func TestBatchMetricProcessor_ReceivingData(t *testing.T) {
 	require.NoError(t, err)
 	require.NoError(t, batcher.Start(context.Background(), componenttest.NewNopHost()))
 
-	metricDataSlice := make([]pdata.Metrics, 0, requestCount)
+	metricDataSlice := make([]pmetric.Metrics, 0, requestCount)
 
 	for requestNum := 0; requestNum < requestCount; requestNum++ {
 		md := testdata.GenerateMetricsManyMetricsSameResource(metricsPerRequest)
@@ -289,7 +290,7 @@ func TestBatchMetricProcessor_ReceivingData(t *testing.T) {
 	}
 
 	// Added to test case with empty resources sent.
-	md := pdata.NewMetrics()
+	md := pmetric.NewMetrics()
 	assert.NoError(t, batcher.ConsumeMetrics(context.Background(), md))
 
 	require.NoError(t, batcher.Shutdown(context.Background()))
@@ -308,7 +309,7 @@ func TestBatchMetricProcessor_ReceivingData(t *testing.T) {
 }
 
 func TestBatchMetricProcessor_BatchSize(t *testing.T) {
-	sizer := otlp.NewProtobufMetricsMarshaler().(pdata.MetricsSizer)
+	sizer := pmetric.NewProtoMarshaler().(pmetric.Sizer)
 	views := MetricViews()
 	require.NoError(t, view.Register(views...))
 	defer view.Unregister(views...)
@@ -470,8 +471,8 @@ func getTestSpanName(requestNum, index int) string {
 	return fmt.Sprintf("test-span-%d-%d", requestNum, index)
 }
 
-func spansReceivedByName(tds []pdata.Traces) map[string]pdata.Span {
-	spansReceivedByName := map[string]pdata.Span{}
+func spansReceivedByName(tds []ptrace.Traces) map[string]ptrace.Span {
+	spansReceivedByName := map[string]ptrace.Span{}
 	for i := range tds {
 		rss := tds[i].ResourceSpans()
 		for i := 0; i < rss.Len(); i++ {
@@ -488,8 +489,8 @@ func spansReceivedByName(tds []pdata.Traces) map[string]pdata.Span {
 	return spansReceivedByName
 }
 
-func metricsReceivedByName(mds []pdata.Metrics) map[string]pdata.Metric {
-	metricsReceivedByName := map[string]pdata.Metric{}
+func metricsReceivedByName(mds []pmetric.Metrics) map[string]pmetric.Metric {
+	metricsReceivedByName := map[string]pmetric.Metric{}
 	for _, md := range mds {
 		rms := md.ResourceMetrics()
 		for i := 0; i < rms.Len(); i++ {
@@ -511,7 +512,7 @@ func getTestMetricName(requestNum, index int) string {
 }
 
 func BenchmarkTraceSizeBytes(b *testing.B) {
-	sizer := otlp.NewProtobufTracesMarshaler().(pdata.TracesSizer)
+	sizer := ptrace.NewProtoMarshaler().(ptrace.Sizer)
 	td := testdata.GenerateTracesManySpansSameResource(8192)
 	for n := 0; n < b.N; n++ {
 		fmt.Println(sizer.TracesSize(td))
@@ -541,7 +542,7 @@ func BenchmarkBatchMetricProcessor(b *testing.B) {
 	require.NoError(b, err)
 	require.NoError(b, batcher.Start(ctx, componenttest.NewNopHost()))
 
-	mds := make([]pdata.Metrics, 0, b.N)
+	mds := make([]pmetric.Metrics, 0, b.N)
 	for n := 0; n < b.N; n++ {
 		mds = append(mds,
 			testdata.GenerateMetricsManyMetricsSameResource(metricsPerRequest),
@@ -567,7 +568,7 @@ func (sme *metricsSink) Capabilities() consumer.Capabilities {
 	}
 }
 
-func (sme *metricsSink) ConsumeMetrics(_ context.Context, md pdata.Metrics) error {
+func (sme *metricsSink) ConsumeMetrics(_ context.Context, md pmetric.Metrics) error {
 	sme.mu.Lock()
 	defer sme.mu.Unlock()
 	sme.metricsCount += md.MetricCount()
@@ -592,7 +593,7 @@ func TestBatchLogProcessor_ReceivingData(t *testing.T) {
 	require.NoError(t, err)
 	require.NoError(t, batcher.Start(context.Background(), componenttest.NewNopHost()))
 
-	logDataSlice := make([]pdata.Logs, 0, requestCount)
+	logDataSlice := make([]plog.Logs, 0, requestCount)
 
 	for requestNum := 0; requestNum < requestCount; requestNum++ {
 		ld := testdata.GenerateLogsManyLogRecordsSameResource(logsPerRequest)
@@ -605,7 +606,7 @@ func TestBatchLogProcessor_ReceivingData(t *testing.T) {
 	}
 
 	// Added to test case with empty resources sent.
-	ld := pdata.NewLogs()
+	ld := plog.NewLogs()
 	assert.NoError(t, batcher.ConsumeLogs(context.Background(), ld))
 
 	require.NoError(t, batcher.Shutdown(context.Background()))
@@ -624,7 +625,7 @@ func TestBatchLogProcessor_ReceivingData(t *testing.T) {
 }
 
 func TestBatchLogProcessor_BatchSize(t *testing.T) {
-	sizer := otlp.NewProtobufLogsMarshaler().(pdata.LogsSizer)
+	sizer := plog.NewProtoMarshaler().(plog.Sizer)
 	views := MetricViews()
 	require.NoError(t, view.Register(views...))
 	defer view.Unregister(views...)
@@ -767,8 +768,8 @@ func getTestLogSeverityText(requestNum, index int) string {
 	return fmt.Sprintf("test-log-int-%d-%d", requestNum, index)
 }
 
-func logsReceivedBySeverityText(lds []pdata.Logs) map[string]pdata.LogRecord {
-	logsReceivedBySeverityText := map[string]pdata.LogRecord{}
+func logsReceivedBySeverityText(lds []plog.Logs) map[string]plog.LogRecord {
+	logsReceivedBySeverityText := map[string]plog.LogRecord{}
 	for i := range lds {
 		ld := lds[i]
 		rms := ld.ResourceLogs()